# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Learning Loop Agent API
  version: 0.2.0
  description: |
    AI 学习工具的目标 HTTP 契约。它覆盖 Ticket 01-16，Ticket 17-21 复用其中定义的工作流契约，
    是后端实现和前端集成的共同输入，
    不表示当前后端已经完成。除文件上传与下载外，请求和响应均使用 application/json。
    v0.2 使用 chat_style 表示无阶段门禁的对话风格，并用 api_format 显式选择模型调用协议。
    旧 learning_mode、SocraticState 和阶段型 intent 只用于迁移期兼容。
  x-contract-status: target
servers:
  - url: http://127.0.0.1:4173
    description: 本地开发服务
security: []
...
paths:
...
  /api/subjects:
    get:
      operationId: listSubjects
      tags: [Subjects]
      summary: 列出科目空间
      x-tickets: ['01']
      responses:
        '200':
          description: 科目空间列表
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/SubjectList
    post:
      operationId: createSubject
      tags: [Subjects]
      summary: 创建并激活科目空间
      x-tickets: ['01']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas.yaml#/components/schemas/SubjectInput
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/Subject
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/subjects/import:
    post:
      operationId: importSubject
      tags: [Subjects]
      summary: 导入科目空间副本
      description: |
        上传由 exportSubject 生成的 zip。保留原科目与资料版本 id。
        若工作区已有相同科目 id 或任一资料版本 id，整体拒绝，不做重映射或合并。
        科目名与现有重名但 id 不同时，导入名自动加「（导入）」后缀。
        先全量校验再落盘；损坏 zip 或缺 manifest 返回 422，不产生半导入。
      x-tickets: ['33']
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: ./schemas.yaml#/components/schemas/SubjectImportUpload
      responses:
        '201':
          description: 已导入科目空间
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/Subject
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/subjects/{subject_id}:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      operationId: getSubject
      tags: [Subjects]
      summary: 读取科目空间
      x-tickets: ['01']
      responses:
        '200':
          description: 科目空间
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/Subject
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      operationId: updateSubject
      tags: [Subjects]
      summary: 重命名科目空间
      x-tickets: ['01']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas.yaml#/components/schemas/SubjectInput
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/Subject
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
    delete:
      operationId: deleteSubject
      tags: [Subjects]
      summary: 删除科目空间
      description: 删除其当前数据；已导出的副本不受影响。存在运行中任务时拒绝删除。
      x-tickets: ['01']
      responses:
        '204': {description: 已删除}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/subjects/{subject_id}/activate:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    post:
      operationId: activateSubject
      tags: [Subjects]
      summary: 激活科目空间
      x-tickets: ['01']
      responses:
        '200':
          description: 已激活
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/Workspace
        '404': {$ref: '#/components/responses/NotFound'}
  /api/subjects/{subject_id}/export:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      operationId: exportSubject
      tags: [Subjects]
      summary: 导出科目空间副本
      description: |
        同步返回 zip：manifest.json、完整科目节点 subject.json、
        该科目资料原文件/解析索引/缓存（sources/），以及属于该科目资料版本的 citations.json。
        不走异步任务。
      x-tickets: ['33']
      responses:
        '200':
          description: 科目空间 zip 副本
          content:
            application/zip:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/models:
    get:
      operationId: listModels
      tags: [Models]
      summary: 列出模型服务
      x-tickets: ['02', '05']
      responses:
        '200':
          description: 模型服务列表
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/ModelServiceList
    post:
      operationId: createModel
      tags: [Models]
      summary: 添加模型服务
      x-tickets: ['02', '05']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas.yaml#/components/schemas/ModelServiceInput
      responses:
        '201':
          description: 已添加；响应不返回 API Key
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/ModelService
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/models/{model_id}:
    parameters:
      - $ref: '#/components/parameters/ModelId'
    get:
      operationId: getModel
      tags: [Models]
      summary: 读取模型服务
      x-tickets: ['02', '05']
      responses:
        '200':
          description: 模型服务；不返回 API Key
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/ModelService
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      operationId: updateModel
      tags: [Models]
      summary: 更新模型服务
      x-tickets: ['02', '05']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas.yaml#/components/schemas/ModelServicePatch
      responses:
        '200':
          description: 已更新；验证状态被重置
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/ModelService
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
    delete:
      operationId: deleteModel
      tags: [Models]
      summary: 删除模型服务
      description: 历史 AI 消息和批改结果中的模型快照保持不变。
      x-tickets: ['02']
      responses:
        '204': {description: 已删除}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/models/{model_id}/verify:
    parameters:
      - $ref: '#/components/parameters/ModelId'
    post:
      operationId: verifyModel
      tags: [Models, Operations]
      summary: 验证模型服务和能力
      description: 验证文本输入，并按配置或能力探测记录是否支持图片；不按服务商名称推断能力。
      x-tickets: ['02', '05']
      responses:
        '202': {$ref: '#/components/responses/OperationAccepted'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/subjects/{subject_id}/chat:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      operationId: getChat
      tags: [Chat]
      summary: 读取完整会话记录
      x-tickets: ['02', '04', '06', '07', '12']
      responses:
        '200':
          description: 当前科目的会话记录
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Chat}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      operationId: configureChat
      tags: [Chat]
      summary: 配置兼容聊天、资料范围和依据模式
      description: chat_style 只改变后续回复的提示风格；旧 learning_mode 为兼容字段。切换配置不删除已有会话记录，也不隐式切换模型。
      x-tickets: ['04', '06', '07']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: ./schemas.yaml#/components/schemas/ChatConfigPatch}
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Chat}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
    delete:
      operationId: clearChat
      tags: [Chat]
      summary: 删除当前科目的会话记录
      x-tickets: ['02']
      responses:
        '204': {description: 已删除会话消息，保留会话配置}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/subjects/{subject_id}/chat/model:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    post:
      operationId: switchChatModel
      tags: [Chat]
      summary: 主动切换当前会话使用的模型
      x-tickets: ['02']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: ./schemas.yaml#/components/schemas/ChatModelInput}
      responses:
        '200':
          description: 已切换；历史消息模型快照不变
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Chat}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/subjects/{subject_id}/chat/messages:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    post:
      operationId: createChatMessage
      tags: [Chat, Operations]
      summary: 发送自然语言消息
      description: |
        chat_style 只改变提示或工具指令，不建立教学阶段或限制用户下一条消息。旧 intent 仅作为兼容字段。
        选区问答把选区、所在题目和来源引用一起提交。
        响应只创建会话消息与异步任务，不会修改右侧试卷或学习产物。
      x-tickets: ['02', '04', '06', '12']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: ./schemas.yaml#/components/schemas/ChatMessageInput}
      responses:
        '202': {$ref: '#/components/responses/OperationAccepted'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/subjects/{subject_id}/chat/crash-course:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    post:
      operationId: generateCrashCourse
      tags: [Chat, Artifacts, Operations]
      summary: 通过兼容接口生成章节速成学习产物
      description: 新客户端使用普通消息的 chat_style=crash-course；只有用户明确保存时才创建 AI 文档。
      deprecated: true
      x-tickets: ['07']
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: ./schemas.yaml#/components/schemas/CrashCourseInput}
      responses:
        '202': {$ref: '#/components/responses/OperationAccepted'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/subjects/{subject_id}/chat/stop:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    post:
      operationId: stopSubjectChatGeneration
      tags: [Chat, Operations]
      summary: 停止当前科目的会话生成
      x-tickets: ['02']
      responses:
        '202': {$ref: '#/components/responses/OperationAccepted'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/generations/{generation_id}/stop:
    parameters:
      - $ref: '#/components/parameters/GenerationId'
    post:
      operationId: stopGeneration
      tags: [Chat, Operations]
      summary: 按兼容任务 ID 停止会话生成
      description: generation_id 与对应异步任务的 operation_id 相同；新客户端优先使用 operations 取消接口。
      deprecated: true
      x-tickets: ['02']
      responses:
        '202': {$ref: '#/components/responses/OperationAccepted'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/subjects/{subject_id}/artifacts:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      operationId: listArtifacts
      tags: [Artifacts]
      summary: 列出学习产物
      x-tickets: ['07']
      responses:
        '200':
          description: 学习产物列表
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/LearningArtifactList}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/artifacts/{artifact_id}:
    parameters:
      - $ref: '#/components/parameters/ArtifactId'
    get:
      operationId: getArtifact
      tags: [Artifacts]
      summary: 读取学习产物
      x-tickets: ['07']
      responses:
        '200':
          description: 学习产物
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/LearningArtifact}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      operationId: deleteArtifact
      tags: [Artifacts]
      summary: 删除学习产物
      x-tickets: ['07']
      responses:
        '204': {description: 已删除；会话记录保留不可用的产物引用}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  parameters:
    SubjectId: {name: subject_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
    ModelId: {name: model_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
...
    ArtifactId: {name: artifact_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
...
    GenerationId: {name: generation_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
...
  responses:
    OperationAccepted:
      description: 已接受异步任务
      content:
        application/json:
          schema: {$ref: ./schemas.yaml#/components/schemas/OperationAccepted}
    NotFound:
      description: 资源不存在或不属于当前科目空间
      content:
        application/json:
          schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}
    Conflict:
      description: 当前资源状态不允许此操作
      content:
        application/json:
          schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}
    ValidationFailed:
      description: 请求字段或领域约束校验失败
      content:
        application/json:
          schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}